chore: actually typecheck deno files, fix errors
We weren't including Deno files in `tsc` typechecking, but we also weren't using any `deno` commands
that typecheck.

This introduces typechecking in CI and fixes resulting hidden errors.

The JSON hack isn't great but really it's the whole "import a json and write new data to it" pattern
that should be simplified to "just read the new data in memory", since there's no actual reason to
write it to disk if you follow the code paths. It was just done that way because it was the shortest
path to add the dynamic fissue annotation from the previous implementation.