chore: multiple fixes to e2e report (#560)
* chore: remove unused deno.json field

When running deno, it warns that this is unused when also specifying `imports`.

* chore: fix type errors in Deno files

We weren't including Deno files in `tsc` typechecking, but we also weren't using any `deno` commands
that typecheck. This fixes obscured errors.

I wanted to introduce typechecking in CI in this commit, but it turned out harder than I expected,
so I've pulled that into a separate PR.

The JSON hack isn't great but really it's the whole "import a json and write new data to it" pattern
that should be simplified to "just read the new data in memory", since there's no actual reason to
write it to disk if you follow the code paths. It was just done that way because it was the shortest
path to add the dynamic fissue annotation from the previous implementation.

* chore(e2e-report): fix handling of skipped suites

Some parts of the code weren't correctly handling the case where an entire suite is skipped, with no
specific skipped tests specified.

1. The `testCount.skipped` before this fix was strangely the number of specifically skipped tests
   plus the total number of suites which are either entirely skipped or contain skipped tests (i.e.
   it was pretty wrong).
2. The computed `skipped` count within each test suite was also wrong but in a different way. It was
   not counting at all skipped tests where the entire suite is skipped.

I've also commented some existing correct logic that is very unintuitive.

* chore(e2e): split open issues and skipped tests

It was really confusing, both as a person viewing the page, and as a developer working with the
code. This code and this component were muddling three different data types: failed tests, skipped
tests, and skipped test suites. It was really error prone and hard to follow and modify.

I've split this into separate components and sections on the page for failed tests ("open issues")
and skipped tests, and I've passed skipped tests and skipped suites separately into the latter
component.

I've also refactored some code with side effects to be easier to work with.

* refactor(e2e): remove stale todo

* chore(e2e): update e2e test-results.json fixture

This is the latest run on "latest" against main, from last night.

* chore: remove unused deno scripts

* chore(e2e): add test-results.json fixture to second path

We expect this in two different locations depending on the code path... Having a sample file here
allows for TypeScript to infer the type correctly.

* chore(e2e): add back some skipped tests we removed

The tests were named on canary, but we still need the old test names for `latest`.

* chore(e2e): fix bad skipped test name, skip 2 more

* chore(e2e): add missing tests to e2e-skip-retry.json

* chore(e2e): add one more skipped test

* chore(e2e): fix appending of skipped suites

It was appending entirely skipped suites AND partially skipped suites.

* chore(e2e): update test-results.json fixtures again

* chore(e2e): add sample issues.json for local dev