chore: use netlify-cli as a regular dev dep
Instead of adding nondeterminism to our CI by installing netlify-cli@latest on each run, just
install it as a regular dev dep and let Renovate handle bumps.

In the case of the next.js repo e2e tests, reference this repo (which is guaranteed to be a sibling
directory) to access the netlify-cli binary.